Move some methods to separate classes.
