+++ src/Bibliotheca.Server.Indexer.Nightcrawler.Core/Services/IDiscoveryService.cs
@@ -1,0 +1,9 @@
+using System.Threading.Tasks;
+
+namespace Bibliotheca.Server.Indexer.Nightcrawler.Core.Services
+{
+    public interface IDiscoveryService
+    {
+        Task<string> GetGatewayAddress();
+    }
+}
